Changed routes as Jeff wanted
Changed old UserHome component to Home component
Downloaded original UserHome component and tests from boilermaker
