+++ client/components/user-home.js
@@ -1,42 +1,31 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import { ProductList } from './product-list';
 
 /**
  * COMPONENT
  */
-export const UserHome = ({ newProducts, saleProducts }) => {
-  return (
+export const UserHome = ({ email }) =>
+  (
     <div>
-      <div id="newProducts">
-        <h1 className="headertext">Hot off the Dark Web</h1>
-        <ProductList products={newProducts} />
-      </div>
-      <div id="saleProducts">
-        <h1 className="headertext">Pennies on the Dollar</h1>
-        <ProductList products={saleProducts} />
-      </div>
+      <h3>Welcome, {email}</h3>
     </div>
   );
-};
 
 /**
  * CONTAINER
  */
-const mapStateToProps = (state) => {
+const mapState = (state) => {
   return {
-    newProducts: state.products.filter(product => product.tags.includes('new')),
-    saleProducts: state.products.filter(product => product.tags.includes('sale')),
+    email: state.user.email,
   };
 };
 
-export default connect(mapStateToProps)(UserHome);
+export default connect(mapState)(UserHome);
 
 /**
  * PROP TYPES
  */
 UserHome.propTypes = {
-  newProducts: PropTypes.array,
-  saleProducts: PropTypes.array,
+  email: PropTypes.string,
 };
