+++ client/routes.js
@@ -4,7 +4,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { withRouter, Route, Switch } from 'react-router-dom';
 import PropTypes from 'prop-types';
-import { Login, Signup, UserHome, ProductList, Product } from './components';
+import { Login, Signup, Home, UserHome, ProductList, Product } from './components';
 import { me, fetchAllProducts } from './store';
 
 /**
@@ -21,6 +21,7 @@ class Routes extends Component {
     return (
       <Switch>
         {/* Routes placed here are available to all visitors */}
+        <Route exact path="/" component={Home} />
         <Route path="/login" component={Login} />
         <Route path="/signup" component={Signup} />
         <Route exact path="/products" component={ProductList} />
@@ -32,8 +33,8 @@ class Routes extends Component {
               <Route path="/home" component={UserHome} />
             </Switch>
         }
-        {/* Displays our Login component as a fallback */}
-        <Route component={Login} />
+        {/* Displays our home component as a fallback */}
+        <Route component={Home} />
       </Switch>
     );
   }
